+++ velo-spring-boot2-autoconfigure/src/main/java/io/github/luminion/velo/redis/VeloRedisAutoConfiguration.java
@@ -1,6 +1,7 @@
 package io.github.luminion.velo.redis;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
 import org.springframework.context.annotation.Import;
 
@@ -11,6 +12,10 @@ import org.springframework.context.annotation.Import;
  * @since 1.3.1
  */
 @AutoConfiguration(before = RedisAutoConfiguration.class)
+@ConditionalOnClass(name = {
+        "org.springframework.data.redis.core.RedisOperations",
+        "org.springframework.data.redis.connection.RedisConnectionFactory"
+})
 @Import({VeloBoot2RedisConnectionConfigurationImportSelector.class, VeloRedisConfiguration.class})
 public class VeloRedisAutoConfiguration {
 }